test: move device_page mocks to test_utils
