+++ test/test_utils.mocks.dart
@@ -7,9 +7,11 @@ import 'dart:async' as _i4;
 import 'dart:io' as _i5;
 import 'dart:ui' as _i7;
 
+import 'package:firmware_updater/device_model.dart' as _i8;
+import 'package:firmware_updater/device_store.dart' as _i9;
 import 'package:firmware_updater/fwupd_notifier.dart' as _i6;
-import 'package:firmware_updater/fwupd_service.dart' as _i2;
-import 'package:fwupd/fwupd.dart' as _i3;
+import 'package:firmware_updater/fwupd_service.dart' as _i3;
+import 'package:fwupd/fwupd.dart' as _i2;
 import 'package:mockito/mockito.dart' as _i1;
 
 // ignore_for_file: type=lint
@@ -23,19 +25,29 @@ import 'package:mockito/mockito.dart' as _i1;
 // ignore_for_file: camel_case_types
 // ignore_for_file: subtype_of_sealed_class
 
+class _FakeFwupdDevice_0 extends _i1.SmartFake implements _i2.FwupdDevice {
+  _FakeFwupdDevice_0(
+    Object parent,
+    Invocation parentInvocation,
+  ) : super(
+          parent,
+          parentInvocation,
+        );
+}
+
 /// A class which mocks [FwupdService].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockFwupdService extends _i1.Mock implements _i2.FwupdService {
+class MockFwupdService extends _i1.Mock implements _i3.FwupdService {
   MockFwupdService() {
     _i1.throwOnMissingStub(this);
   }
 
   @override
-  _i3.FwupdStatus get status => (super.noSuchMethod(
+  _i2.FwupdStatus get status => (super.noSuchMethod(
         Invocation.getter(#status),
-        returnValue: _i3.FwupdStatus.unknown,
-      ) as _i3.FwupdStatus);
+        returnValue: _i2.FwupdStatus.unknown,
+      ) as _i2.FwupdStatus);
   @override
   int get percentage => (super.noSuchMethod(
         Invocation.getter(#percentage),
@@ -47,25 +59,25 @@ class MockFwupdService extends _i1.Mock implements _i2.FwupdService {
         returnValue: '',
       ) as String);
   @override
-  _i4.Stream<_i3.FwupdDevice> get deviceAdded => (super.noSuchMethod(
+  _i4.Stream<_i2.FwupdDevice> get deviceAdded => (super.noSuchMethod(
         Invocation.getter(#deviceAdded),
-        returnValue: _i4.Stream<_i3.FwupdDevice>.empty(),
-      ) as _i4.Stream<_i3.FwupdDevice>);
+        returnValue: _i4.Stream<_i2.FwupdDevice>.empty(),
+      ) as _i4.Stream<_i2.FwupdDevice>);
   @override
-  _i4.Stream<_i3.FwupdDevice> get deviceChanged => (super.noSuchMethod(
+  _i4.Stream<_i2.FwupdDevice> get deviceChanged => (super.noSuchMethod(
         Invocation.getter(#deviceChanged),
-        returnValue: _i4.Stream<_i3.FwupdDevice>.empty(),
-      ) as _i4.Stream<_i3.FwupdDevice>);
+        returnValue: _i4.Stream<_i2.FwupdDevice>.empty(),
+      ) as _i4.Stream<_i2.FwupdDevice>);
   @override
-  _i4.Stream<_i3.FwupdDevice> get deviceRemoved => (super.noSuchMethod(
+  _i4.Stream<_i2.FwupdDevice> get deviceRemoved => (super.noSuchMethod(
         Invocation.getter(#deviceRemoved),
-        returnValue: _i4.Stream<_i3.FwupdDevice>.empty(),
-      ) as _i4.Stream<_i3.FwupdDevice>);
+        returnValue: _i4.Stream<_i2.FwupdDevice>.empty(),
+      ) as _i4.Stream<_i2.FwupdDevice>);
   @override
-  _i4.Stream<_i3.FwupdDevice> get deviceRequest => (super.noSuchMethod(
+  _i4.Stream<_i2.FwupdDevice> get deviceRequest => (super.noSuchMethod(
         Invocation.getter(#deviceRequest),
-        returnValue: _i4.Stream<_i3.FwupdDevice>.empty(),
-      ) as _i4.Stream<_i3.FwupdDevice>);
+        returnValue: _i4.Stream<_i2.FwupdDevice>.empty(),
+      ) as _i4.Stream<_i2.FwupdDevice>);
   @override
   _i4.Stream<List<String>> get propertiesChanged => (super.noSuchMethod(
         Invocation.getter(#propertiesChanged),
@@ -123,7 +135,7 @@ class MockFwupdService extends _i1.Mock implements _i2.FwupdService {
         returnValueForMissingStub: _i4.Future<void>.value(),
       ) as _i4.Future<void>);
   @override
-  _i4.Future<void> activate(_i3.FwupdDevice? device) => (super.noSuchMethod(
+  _i4.Future<void> activate(_i2.FwupdDevice? device) => (super.noSuchMethod(
         Invocation.method(
           #activate,
           [device],
@@ -132,7 +144,7 @@ class MockFwupdService extends _i1.Mock implements _i2.FwupdService {
         returnValueForMissingStub: _i4.Future<void>.value(),
       ) as _i4.Future<void>);
   @override
-  _i4.Future<void> clearResults(_i3.FwupdDevice? device) => (super.noSuchMethod(
+  _i4.Future<void> clearResults(_i2.FwupdDevice? device) => (super.noSuchMethod(
         Invocation.method(
           #clearResults,
           [device],
@@ -141,66 +153,66 @@ class MockFwupdService extends _i1.Mock implements _i2.FwupdService {
         returnValueForMissingStub: _i4.Future<void>.value(),
       ) as _i4.Future<void>);
   @override
-  _i4.Future<List<_i3.FwupdDevice>> getDevices() => (super.noSuchMethod(
+  _i4.Future<List<_i2.FwupdDevice>> getDevices() => (super.noSuchMethod(
         Invocation.method(
           #getDevices,
           [],
         ),
         returnValue:
-            _i4.Future<List<_i3.FwupdDevice>>.value(<_i3.FwupdDevice>[]),
-      ) as _i4.Future<List<_i3.FwupdDevice>>);
+            _i4.Future<List<_i2.FwupdDevice>>.value(<_i2.FwupdDevice>[]),
+      ) as _i4.Future<List<_i2.FwupdDevice>>);
   @override
-  _i4.Future<List<_i3.FwupdRelease>> getDowngrades(_i3.FwupdDevice? device) =>
+  _i4.Future<List<_i2.FwupdRelease>> getDowngrades(_i2.FwupdDevice? device) =>
       (super.noSuchMethod(
         Invocation.method(
           #getDowngrades,
           [device],
         ),
         returnValue:
-            _i4.Future<List<_i3.FwupdRelease>>.value(<_i3.FwupdRelease>[]),
-      ) as _i4.Future<List<_i3.FwupdRelease>>);
+            _i4.Future<List<_i2.FwupdRelease>>.value(<_i2.FwupdRelease>[]),
+      ) as _i4.Future<List<_i2.FwupdRelease>>);
   @override
-  _i4.Future<List<_i3.FwupdPlugin>> getPlugins() => (super.noSuchMethod(
+  _i4.Future<List<_i2.FwupdPlugin>> getPlugins() => (super.noSuchMethod(
         Invocation.method(
           #getPlugins,
           [],
         ),
         returnValue:
-            _i4.Future<List<_i3.FwupdPlugin>>.value(<_i3.FwupdPlugin>[]),
-      ) as _i4.Future<List<_i3.FwupdPlugin>>);
+            _i4.Future<List<_i2.FwupdPlugin>>.value(<_i2.FwupdPlugin>[]),
+      ) as _i4.Future<List<_i2.FwupdPlugin>>);
   @override
-  _i4.Future<List<_i3.FwupdRelease>> getReleases(_i3.FwupdDevice? device) =>
+  _i4.Future<List<_i2.FwupdRelease>> getReleases(_i2.FwupdDevice? device) =>
       (super.noSuchMethod(
         Invocation.method(
           #getReleases,
           [device],
         ),
         returnValue:
-            _i4.Future<List<_i3.FwupdRelease>>.value(<_i3.FwupdRelease>[]),
-      ) as _i4.Future<List<_i3.FwupdRelease>>);
+            _i4.Future<List<_i2.FwupdRelease>>.value(<_i2.FwupdRelease>[]),
+      ) as _i4.Future<List<_i2.FwupdRelease>>);
   @override
-  _i4.Future<List<_i3.FwupdRemote>> getRemotes() => (super.noSuchMethod(
+  _i4.Future<List<_i2.FwupdRemote>> getRemotes() => (super.noSuchMethod(
         Invocation.method(
           #getRemotes,
           [],
         ),
         returnValue:
-            _i4.Future<List<_i3.FwupdRemote>>.value(<_i3.FwupdRemote>[]),
-      ) as _i4.Future<List<_i3.FwupdRemote>>);
+            _i4.Future<List<_i2.FwupdRemote>>.value(<_i2.FwupdRemote>[]),
+      ) as _i4.Future<List<_i2.FwupdRemote>>);
   @override
-  _i4.Future<List<_i3.FwupdRelease>> getUpgrades(_i3.FwupdDevice? device) =>
+  _i4.Future<List<_i2.FwupdRelease>> getUpgrades(_i2.FwupdDevice? device) =>
       (super.noSuchMethod(
         Invocation.method(
           #getUpgrades,
           [device],
         ),
         returnValue:
-            _i4.Future<List<_i3.FwupdRelease>>.value(<_i3.FwupdRelease>[]),
-      ) as _i4.Future<List<_i3.FwupdRelease>>);
+            _i4.Future<List<_i2.FwupdRelease>>.value(<_i2.FwupdRelease>[]),
+      ) as _i4.Future<List<_i2.FwupdRelease>>);
   @override
   _i4.Future<void> install(
-    _i3.FwupdDevice? device,
-    _i3.FwupdRelease? release, [
+    _i2.FwupdDevice? device,
+    _i2.FwupdRelease? release, [
     _i5.ResourceHandle Function(_i5.RandomAccessFile)? resourceHandleFromFile,
   ]) =>
       (super.noSuchMethod(
@@ -216,7 +228,7 @@ class MockFwupdService extends _i1.Mock implements _i2.FwupdService {
         returnValueForMissingStub: _i4.Future<void>.value(),
       ) as _i4.Future<void>);
   @override
-  _i4.Future<void> unlock(_i3.FwupdDevice? device) => (super.noSuchMethod(
+  _i4.Future<void> unlock(_i2.FwupdDevice? device) => (super.noSuchMethod(
         Invocation.method(
           #unlock,
           [device],
@@ -225,7 +237,7 @@ class MockFwupdService extends _i1.Mock implements _i2.FwupdService {
         returnValueForMissingStub: _i4.Future<void>.value(),
       ) as _i4.Future<void>);
   @override
-  _i4.Future<void> verify(_i3.FwupdDevice? device) => (super.noSuchMethod(
+  _i4.Future<void> verify(_i2.FwupdDevice? device) => (super.noSuchMethod(
         Invocation.method(
           #verify,
           [device],
@@ -234,7 +246,7 @@ class MockFwupdService extends _i1.Mock implements _i2.FwupdService {
         returnValueForMissingStub: _i4.Future<void>.value(),
       ) as _i4.Future<void>);
   @override
-  _i4.Future<void> verifyUpdate(_i3.FwupdDevice? device) => (super.noSuchMethod(
+  _i4.Future<void> verifyUpdate(_i2.FwupdDevice? device) => (super.noSuchMethod(
         Invocation.method(
           #verifyUpdate,
           [device],
@@ -262,10 +274,10 @@ class MockFwupdNotifier extends _i1.Mock implements _i6.FwupdNotifier {
   }
 
   @override
-  _i3.FwupdStatus get status => (super.noSuchMethod(
+  _i2.FwupdStatus get status => (super.noSuchMethod(
         Invocation.getter(#status),
-        returnValue: _i3.FwupdStatus.unknown,
-      ) as _i3.FwupdStatus);
+        returnValue: _i2.FwupdStatus.unknown,
+      ) as _i2.FwupdStatus);
   @override
   int get percentage => (super.noSuchMethod(
         Invocation.getter(#percentage),
@@ -321,7 +333,7 @@ class MockFwupdNotifier extends _i1.Mock implements _i6.FwupdNotifier {
       );
   @override
   void registerDeviceRequestListener(
-          dynamic Function(_i3.FwupdDevice)? deviceRequestListener) =>
+          dynamic Function(_i2.FwupdDevice)? deviceRequestListener) =>
       super.noSuchMethod(
         Invocation.method(
           #registerDeviceRequestListener,
@@ -329,6 +341,222 @@ class MockFwupdNotifier extends _i1.Mock implements _i6.FwupdNotifier {
         ),
         returnValueForMissingStub: null,
       );
+  @override
+  _i4.Future<void> refresh() => (super.noSuchMethod(
+        Invocation.method(
+          #refresh,
+          [],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+  @override
+  _i4.Future<void> dispose() => (super.noSuchMethod(
+        Invocation.method(
+          #dispose,
+          [],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+  @override
+  void addListener(_i7.VoidCallback? listener) => super.noSuchMethod(
+        Invocation.method(
+          #addListener,
+          [listener],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  void removeListener(_i7.VoidCallback? listener) => super.noSuchMethod(
+        Invocation.method(
+          #removeListener,
+          [listener],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  void notifyListeners() => super.noSuchMethod(
+        Invocation.method(
+          #notifyListeners,
+          [],
+        ),
+        returnValueForMissingStub: null,
+      );
+}
+
+/// A class which mocks [DeviceModel].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockDeviceModel extends _i1.Mock implements _i8.DeviceModel {
+  MockDeviceModel() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  _i2.FwupdDevice get device => (super.noSuchMethod(
+        Invocation.getter(#device),
+        returnValue: _FakeFwupdDevice_0(
+          this,
+          Invocation.getter(#device),
+        ),
+      ) as _i2.FwupdDevice);
+  @override
+  bool get onBattery => (super.noSuchMethod(
+        Invocation.getter(#onBattery),
+        returnValue: false,
+      ) as bool);
+  @override
+  bool get hasListeners => (super.noSuchMethod(
+        Invocation.getter(#hasListeners),
+        returnValue: false,
+      ) as bool);
+  @override
+  bool get isDisposed => (super.noSuchMethod(
+        Invocation.getter(#isDisposed),
+        returnValue: false,
+      ) as bool);
+  @override
+  _i4.Future<void> init() => (super.noSuchMethod(
+        Invocation.method(
+          #init,
+          [],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+  @override
+  _i4.Future<void> dispose() => (super.noSuchMethod(
+        Invocation.method(
+          #dispose,
+          [],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+  @override
+  _i4.Future<void> update(_i2.FwupdDevice? device) => (super.noSuchMethod(
+        Invocation.method(
+          #update,
+          [device],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+  @override
+  _i4.Future<void> verify() => (super.noSuchMethod(
+        Invocation.method(
+          #verify,
+          [],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+  @override
+  _i4.Future<void> verifyUpdate() => (super.noSuchMethod(
+        Invocation.method(
+          #verifyUpdate,
+          [],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+  @override
+  _i4.Future<void> install(_i2.FwupdRelease? release) => (super.noSuchMethod(
+        Invocation.method(
+          #install,
+          [release],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+  @override
+  bool hasUpgrade() => (super.noSuchMethod(
+        Invocation.method(
+          #hasUpgrade,
+          [],
+        ),
+        returnValue: false,
+      ) as bool);
+  @override
+  void addListener(_i7.VoidCallback? listener) => super.noSuchMethod(
+        Invocation.method(
+          #addListener,
+          [listener],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  void removeListener(_i7.VoidCallback? listener) => super.noSuchMethod(
+        Invocation.method(
+          #removeListener,
+          [listener],
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  void notifyListeners() => super.noSuchMethod(
+        Invocation.method(
+          #notifyListeners,
+          [],
+        ),
+        returnValueForMissingStub: null,
+      );
+}
+
+/// A class which mocks [DeviceStore].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockDeviceStore extends _i1.Mock implements _i9.DeviceStore {
+  MockDeviceStore() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  bool get showReleases => (super.noSuchMethod(
+        Invocation.getter(#showReleases),
+        returnValue: false,
+      ) as bool);
+  @override
+  set showReleases(bool? value) => super.noSuchMethod(
+        Invocation.setter(
+          #showReleases,
+          value,
+        ),
+        returnValueForMissingStub: null,
+      );
+  @override
+  List<_i2.FwupdDevice> get devices => (super.noSuchMethod(
+        Invocation.getter(#devices),
+        returnValue: <_i2.FwupdDevice>[],
+      ) as List<_i2.FwupdDevice>);
+  @override
+  bool get hasListeners => (super.noSuchMethod(
+        Invocation.getter(#hasListeners),
+        returnValue: false,
+      ) as bool);
+  @override
+  bool get isDisposed => (super.noSuchMethod(
+        Invocation.getter(#isDisposed),
+        returnValue: false,
+      ) as bool);
+  @override
+  int indexOf(String? deviceId) => (super.noSuchMethod(
+        Invocation.method(
+          #indexOf,
+          [deviceId],
+        ),
+        returnValue: 0,
+      ) as int);
+  @override
+  _i4.Future<void> init() => (super.noSuchMethod(
+        Invocation.method(
+          #init,
+          [],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
   @override
   _i4.Future<void> refresh() => (super.noSuchMethod(
         Invocation.method(
